+++ src/Validator/SchemaValidator/AllOfValidator.php
@@ -9,7 +9,6 @@ use Duyler\OpenApi\Validator\Error\ValidationContext;
 use Duyler\OpenApi\Validator\Exception\ValidationException;
 use Override;
 
-use function count;
 use function sprintf;
 
 final readonly class AllOfValidator extends AbstractCompositionalValidator implements KeywordApplicable
@@ -33,18 +32,10 @@ final readonly class AllOfValidator extends AbstractCompositionalValidator imple
 
         $result = $this->validateSchemas($schema->allOf, $data, $context, 'allOf');
 
-        if ([] !== $result->errors || [] !== $result->abstractErrors) {
-            $allErrors = $result->abstractErrors;
-
-            foreach ($result->errors as $exception) {
-                foreach ($exception->getErrors() as $error) {
-                    $allErrors[] = $error;
-                }
-            }
-
+        if (0 !== $result->failedCount) {
             throw new ValidationException(
-                sprintf('All of the schemas must match, but %d failed', count($result->errors)),
-                errors: $allErrors,
+                sprintf('All of the schemas must match, but %d failed', $result->failedCount),
+                errors: $result->errors,
             );
         }
     }