+++ src/Validator/SchemaValidator/ValidationResult.php
@@ -5,15 +5,13 @@ declare(strict_types=1);
 namespace Duyler\OpenApi\Validator\SchemaValidator;
 
 use Duyler\OpenApi\Validator\Exception\ValidationErrorInterface;
-use Duyler\OpenApi\Validator\Exception\ValidationException;
 
 final readonly class ValidationResult
 {
     public function __construct(
         public readonly int $validCount,
-        /** @var array<int, ValidationException> */
-        public readonly array $errors,
         /** @var array<int, ValidationErrorInterface> */
-        public readonly array $abstractErrors,
+        public readonly array $errors,
+        public readonly int $failedCount = 0,
     ) {}
 }